requirements.txt
# I think poetry is better than this 'requirements.txt', but its only my opinion

# Django
Django==4.0

# PostgreSQL
psycopg2==2.9.2

# Telegram
python-telegram-bot==13.9

# another depencies
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.10.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1